requirements.txt
async-timeout==3.0.1
discord.py==1.3.3
python-dateutil==2.8.1
python-dotenv==0.13.0
requests==2.23.0